Abstract out commit header as primitive
