+++ ui/DesignSystem/Primitive/Header/index.js
@@ -1,5 +1,7 @@
+import Back from "./Back.svelte";
 import Large from "./Large.svelte";
 
 export default {
+  Back,
   Large,
 };
